Removed Language module
Since reorgnizing the classes in the Language module it's no longer necessary to keep them grouped in a module.
