--- lib/eson/language/language.rb
@@ -1,8 +1,0 @@
-require_relative './rule.rb'
-require_relative './rule_seq.rb'
-
-module Eson
-  module Language
-    
-  end
-end
