+++ test/abstract_syntax_tree_test.rb
@@ -1,27 +1,27 @@
 require 'minitest'
 require 'minitest/autorun'
 require 'minitest/pride'
-require_relative '../lib/eson/language/rule.rb'
+require_relative '../lib/eson/rule.rb'
 
-describe Eson::Language::Rule::AbstractSyntaxTree do
+describe Eson::Rule::AbstractSyntaxTree do
 
   before do
-    @terminal_rule = Eson::Language::Rule.new_terminal_rule(:terminal,
+    @terminal_rule = Eson::Rule.new_terminal_rule(:terminal,
                                                             /rule/)
-    @nonterminal_rule = Eson::Language::Rule.new(:nonterminal,
+    @nonterminal_rule = Eson::Rule.new(:nonterminal,
                                                  /rule/,
                                                  false,
                                                  ["test"])
     @token = @terminal_rule.make_token(:var)
   end
 
-  subject {Eson::Language::Rule::AbstractSyntaxTree}
-  let(:tree) {Eson::Language::Rule::AbstractSyntaxTree::Tree}
+  subject {Eson::Rule::AbstractSyntaxTree}
+  let(:tree) {Eson::Rule::AbstractSyntaxTree::Tree}
   
   describe "create_ast" do
     it "incorrect parameter type" do
       proc {subject.new("error_type")}.
-        must_raise Eson::Language::Rule::AbstractSyntaxTree::InitializationError
+        must_raise Eson::Rule::AbstractSyntaxTree::InitializationError
     end
     describe "empty" do
       before do
@@ -48,7 +48,7 @@ describe Eson::Language::Rule::AbstractSyntaxTree do
     describe "terminal_rule" do
       it "incorrect parameter type" do
         proc {subject.new(@terminal_rule)}.
-          must_raise Eson::Language::Rule::AbstractSyntaxTree::InitializationError
+          must_raise Eson::Rule::AbstractSyntaxTree::InitializationError
       end
     end
     describe "nonterminal_rule" do
@@ -77,7 +77,7 @@ describe Eson::Language::Rule::AbstractSyntaxTree do
     end
     it "node is invalid type" do
       proc {@tree.insert("foo")}
-        .must_raise Eson::Language::Rule::AbstractSyntaxTree::InsertionError
+        .must_raise Eson::Rule::AbstractSyntaxTree::InsertionError
     end
     it "inserted token is leaf of active node" do
       @tree.insert(@token)
@@ -97,7 +97,7 @@ describe Eson::Language::Rule::AbstractSyntaxTree do
     it "fails on closed tree" do
       @tree.close_active
       proc {@tree.insert(@rule)}
-        .must_raise Eson::Language::Rule::AbstractSyntaxTree::ClosedTreeError
+        .must_raise Eson::Rule::AbstractSyntaxTree::ClosedTreeError
     end
     describe "empty_tree" do
       before do
@@ -110,7 +110,7 @@ describe Eson::Language::Rule::AbstractSyntaxTree do
       end
       it "root insertion failed" do
         proc {@empty_tree.insert("error_string")}
-          .must_raise Eson::Language::Rule::AbstractSyntaxTree::InsertionError
+          .must_raise Eson::Rule::AbstractSyntaxTree::InsertionError
       end
     end
   end
