+++ test/syntax_pass_test.rb
@@ -12,7 +12,7 @@ describe Eson::SyntaxPass do
   describe "valid_token_seq" do
     it "creates_tree" do
       tree = get_ast
-      tree.must_be_instance_of Eson::Language::Rule::AbstractSyntaxTree
+      tree.must_be_instance_of Eson::Rule::AbstractSyntaxTree
       tree.closed?.must_equal true
     end
   end
